// client/src/app/_models/spotForm.ts
export interface SpotForm {
    id: number;
    spotName: string;
    latitude: number;
    longitude: number;
    country: string;
    continent: string;
    description: string;
    isBeginner: boolean;
    isAdvanced: boolean;
    january: boolean;
    february: boolean;
    march: boolean;
    april: boolean;
    may: boolean;
    june: boolean;
    july: boolean;
    august: boolean;
    september: boolean;
    october: boolean;
    november: boolean;
    december: boolean;
}